+++ source/GameInterface/Services/MobileParties/Patches/OnPartyInteractionPatch.cs
@@ -1,5 +1,6 @@
 using Common;
 using Common.Logging;
+using GameInterface.Services.MobileParties.Extensions;
 using GameInterface.Services.MobileParties.Handlers;
 using HarmonyLib;
 using Serilog;
@@ -29,8 +30,13 @@ internal class OnPartyInteractionPatch
             return false;
         }
 
-        if (ContainerProvider.TryResolve<OnPartyInteractionHandler>(out var handler) &&
-            handler.TryHandleReciprocalPlayerInteraction(targetParty, engagingParty))
+        if (!CanHandleReciprocalPlayerInteraction(targetParty, engagingParty)) return true;
+        if (!IsReciprocalPlayerInteractionReady(targetParty, engagingParty)) return true;
+        if (!TryGetPartyBases(targetParty, engagingParty, out var targetPartyBase, out var engagingPartyBase))
+            return true;
+
+        if (ContainerProvider.TryResolve<PlayerPartyInteractionHandler>(out var handler) &&
+            handler.TryHandleReciprocalPlayerInteraction(targetPartyBase, engagingPartyBase))
             return false;
 
         return true;
@@ -42,5 +48,39 @@ internal class OnPartyInteractionPatch
             return targetParty.AttachedTo;
 
         return targetParty;
+    }
+
+    internal static bool CanHandleReciprocalPlayerInteraction(MobileParty targetParty, MobileParty engagingParty)
+    {
+        if (ModInformation.IsServer) return false;
+        if (targetParty == null || engagingParty == null) return false;
+        if (targetParty == engagingParty) return false;
+        if (!engagingParty.IsMainParty) return false;
+        if (!engagingParty.IsControlledByThisInstance()) return false;
+        if (!targetParty.IsPlayerParty()) return false;
+
+        return true;
+    }
+
+    internal static bool IsReciprocalPlayerInteractionReady(MobileParty targetParty, MobileParty engagingParty)
+    {
+        if (targetParty.CurrentSettlement != null) return false;
+        if (targetParty.MapEvent != null || engagingParty.MapEvent != null) return false;
+        if (!targetParty.IsEngaging) return false;
+        if (targetParty.ShortTermTargetParty != engagingParty) return false;
+
+        return true;
+    }
+
+    internal static bool TryGetPartyBases(
+        MobileParty targetParty,
+        MobileParty engagingParty,
+        out PartyBase targetPartyBase,
+        out PartyBase engagingPartyBase)
+    {
+        targetPartyBase = targetParty.Party;
+        engagingPartyBase = engagingParty.Party;
+
+        return targetPartyBase != null && engagingPartyBase != null;
     }
 }
